# student-api-express/swagger.yaml
openapi: 3.0.0
info:
  title: Student Management API
  description: A simple RESTful API for managing students.
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local Express server
paths:
  /students:
    get:
      summary: Get a list of all students
      responses:
        '200':
          description: A JSON array of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
    post:
      summary: Create a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentInput'
      responses:
        '201':
          description: Student created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /students/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
        description: The student ID (MongoDB ObjectId)
    get:
      summary: Get a student by ID
      responses:
        '200':
          description: The student data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '404':
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentInput'
      responses:
        '200':
          description: Student updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Partially update a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentInput'
      responses:
        '200':
          description: Student updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a student
      responses:
        '200':
          description: Student deleted successfully
        '404':
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Student:
      type: object
      required:
        - id
        - name
        - email
        - course
        - semester
      properties:
        id:
          type: string
          example: "66fabc1234567890abcdef12"
        name:
          type: string
          example: "Aarav Patel"
        email:
          type: string
          example: "aarav@example.com"
        course:
          type: string
          example: "Computer Science"
        semester:
          type: integer
          example: 5
    StudentInput:
      type: object
      required:
        - name
        - email
        - course
        - semester
      properties:
        name:
          type: string
          example: "Aarav Patel"
        email:
          type: string
          example: "aarav@example.com"
        course:
          type: string
          example: "Computer Science"
        semester:
          type: integer
          example: 5
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Invalid student id"